Combine about and subscribe pages, add fourth button to bottom bar on mobile, fix styles, test
